Break up qubit operations file (#1467)
* breaking up qubit.py

* fixing imports

* rename file to keep import paths

* changelog, eliminate unused imports

* fix changelog, remove unused imports

* black

* really remove unused imports this time

* remove old file, fix imports

* fix tests and black

* black

* fully merge in pr 1364

* fix imports

* update toffoli controls

* phase shift op

* move observables, add submodule docs

* fix imports

* black

* Update .github/CHANGELOG.md


* move some things around, more docs

* oops

* other oops

* move ising gates to parametric

* oops

* fix imports

* more fix imports

* docstring, arithmetic spelling

* Update .github/CHANGELOG.md

* Apply suggestions from code review


* review feedback

* black

* fix adjoint error issues

* minor wording

* Apply suggestions from code review


* black

* Apply suggestions from code review
