+++ .github/CHANGELOG.md
@@ -335,6 +335,9 @@
 
 <h3>Improvements</h3>
 
+* The `./pennylane/ops/qubit.py` file is broken up into a folder of six separate files.
+  [(#1467)](https://github.com/PennyLaneAI/pennylane/pull/1467)
+
 * Changed to using commas as the separator of wires in the string
   representation of `qml.Hamiltonian` objects for multi-qubit terms.
   [(#1465)](https://github.com/PennyLaneAI/pennylane/pull/1465)
