+++ pennylane/ops/__init__.py
@@ -37,6 +37,12 @@ from .qubit import obs as _qubit__obs__
 from .channel import __all__ as _channel__ops__
 
 
+class AdjointError(Exception):
+    """Exception for non-adjointable operations."""
+
+    pass
+
+
 class Identity(CVObservable):
     r"""pennylane.Identity(wires)
     The identity observable :math:`\I`.
