+++ pennylane/ops/qubit/state_preparation.py
@@ -1,0 +1,91 @@
+# Copyright 2018-2021 Xanadu Quantum Technologies Inc.
+
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+
+#     http://www.apache.org/licenses/LICENSE-2.0
+
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""
+This submodule contains the discrete-variable quantum operations concerned
+with preparing a certain state on the device.
+"""
+# pylint:disable=abstract-method,arguments-differ,protected-access
+
+import pennylane as qml
+from pennylane.operation import AnyWires, Operation
+from pennylane.templates.state_preparations import BasisStatePreparation, MottonenStatePreparation
+
+
+class BasisState(Operation):
+    r"""BasisState(n, wires)
+    Prepares a single computational basis state.
+
+    **Details:**
+
+    * Number of wires: Any (the operation can act on any number of wires)
+    * Number of parameters: 1
+    * Gradient recipe: None (integer parameters not supported)
+
+    .. note::
+
+        If the ``BasisState`` operation is not supported natively on the
+        target device, PennyLane will attempt to decompose the operation
+        into :class:`~.PauliX` operations.
+
+    Args:
+        n (array): prepares the basis state :math:`\ket{n}`, where ``n`` is an
+            array of integers from the set :math:`\{0, 1\}`, i.e.,
+            if ``n = np.array([0, 1, 0])``, prepares the state :math:`|010\rangle`.
+        wires (Sequence[int] or int): the wire(s) the operation acts on
+    """
+    num_params = 1
+    num_wires = AnyWires
+    par_domain = "A"
+    grad_method = None
+
+    @staticmethod
+    def decomposition(n, wires):
+        return BasisStatePreparation(n, wires)
+
+    def adjoint(self):
+        raise qml.ops.AdjointError("No adjoint exists for BasisState operations.")
+
+
+class QubitStateVector(Operation):
+    r"""QubitStateVector(state, wires)
+    Prepare subsystems using the given ket vector in the computational basis.
+
+    **Details:**
+
+    * Number of wires: Any (the operation can act on any number of wires)
+    * Number of parameters: 1
+    * Gradient recipe: None
+
+    .. note::
+
+        If the ``QubitStateVector`` operation is not supported natively on the
+        target device, PennyLane will attempt to decompose the operation
+        using the method developed by Möttönen et al. (Quantum Info. Comput.,
+        2005).
+
+    Args:
+        state (array[complex]): a state vector of size 2**len(wires)
+        wires (Sequence[int] or int): the wire(s) the operation acts on
+    """
+    num_params = 1
+    num_wires = AnyWires
+    par_domain = "A"
+    grad_method = None
+
+    @staticmethod
+    def decomposition(state, wires):
+        return MottonenStatePreparation(state, wires)
+
+    def adjoint(self):
+        raise qml.ops.AdjointError("No adjoint exists for QubitStateVector operations.")
